% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-Rdta.R
\docType{package}
\name{Rdta}
\alias{Rdta-package}
\title{Data Transforming Augmentation for Linear Mixed Models}
\description{
The R package \pkg{Rdta} provides a toolbox to fit univariate and multivariate linear mixed models via data transforming augmentation. Users can also fit these models via typical data augmentation for a comparison. It returns either maximum likelihood estimates of unknown model parameters (hyper-parameters) via an EM algorithm or posterior samples of those parameters via a Markov chain Monte Carlo method.
}
\details{
\tabular{ll}{
Package: \tab Rdta\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2024-1-26\cr
License: \tab GPL-2\cr
Main functions: \tab \code{\link{lmm}}\cr
}
}
\references{
Tak, You, Ghosh, Su, Kelly (2019), "Data Transforming Augmentation for Heteroscedastic Models" <\doi{10.1080/10618600.2019.1704295}>
}
\author{
Hyungsuk Tak (maintainer), Kisung You, Sujit K. Ghosh, and Bingyue Su
}
